[CDP-6] IHM states can be cycled in CPU engine

+++ cdp/cdp/src/engine/Indexer.h
@@ -3,26 +3,26 @@
 #include "Transform.h"
 
 struct Indexer {
-	static __device__ int FlatIndex2(int x, int y, int x_max) {
-		int index = x + (y * x_max);
+	static __device__ unsigned long long FlatIndex2(unsigned long long x, unsigned long long y, unsigned long long x_max) {
+		unsigned long long index = x + (y * x_max);
 
 		return index;
 	}
 
-	static __device__ int FlatIndex3(int x, int y, int z, int x_max, int y_max) {
-		int index = x + (y * x_max) + (z * x_max * y_max);
+	static __device__ unsigned long long FlatIndex3(unsigned long long x, unsigned long long y, unsigned long long z, unsigned long long x_max, unsigned long long y_max) {
+		unsigned long long index = x + (y * x_max) + (z * x_max * y_max);
 
 		return index;
 	}
 
-	static __device__ int FlatIndex4(int x, int y, int z, int w, int x_max, int y_max, int z_max) {
-		int index = x + (y * x_max) + (z * x_max * y_max) + (w * x_max * y_max * z_max);
+	static __device__ unsigned long long FlatIndex4(unsigned long long x, unsigned long long y, unsigned long long z, unsigned long long w, unsigned long long x_max, unsigned long long y_max, unsigned long long z_max) {
+		unsigned long long index = x + (y * x_max) + (z * x_max * y_max) + (w * x_max * y_max * z_max);
 
 		return index;
 	}
 	
-	static __device__ Vector3 InverseFlatIndex3(int index, int x_max, int y_max) {
-		int xy_progress = index % (x_max * y_max);
+	static __device__ Vector3 InverseFlatIndex3(unsigned long long index, unsigned long long x_max, unsigned long long y_max) {
+		unsigned long long xy_progress = index % (x_max * y_max);
 
 		Vector3 result{
 			xy_progress % x_max,
@@ -33,9 +33,9 @@ struct Indexer {
 		return result;
 	}
 	
-	static __device__ Vector4 InverseFlatIndex4(int index, int x_max, int y_max, int z_max) {
-		int xyz_progress = index % (x_max * y_max * z_max);
-		int xy_progress = xyz_progress % (x_max * y_max);
+	static __device__ Vector4 InverseFlatIndex4(unsigned long long index, unsigned long long x_max, unsigned long long y_max, unsigned long long z_max) {
+		unsigned long long xyz_progress = index % (x_max * y_max * z_max);
+		unsigned long long xy_progress = xyz_progress % (x_max * y_max);
 
 		Vector4 result{
 			xy_progress % x_max,
